move enhanced uptime card into uptime panel
